Move stitches.config to top level
This is probably more 'standard' also removes the use of the path.
This has been done so the compiled types correctly reference the
compiled types for stitches.config. Without the paths, it worked but
the import was not deferenced which might cause issues when used.
