+++ src/components/AppBar/AppBar.tsx
@@ -1,7 +1,7 @@
 import type * as Polymorphic from '@radix-ui/react-polymorphic'
 import React, { ComponentProps, forwardRef, PropsWithChildren } from 'react'
-import type { CSSProps } from 'stitches.config'
-import { CSS, styled } from 'stitches.config'
+import type { CSSProps } from '../../stitches.config'
+import { CSS, styled } from '../../stitches.config'
 import { Box } from '../Box'
 import { Button } from '../Button'
 import { Heading } from '../Heading'
