+++ src/components/Box/Box.tsx
@@ -1,4 +1,4 @@
-import { styled } from 'stitches.config'
+import { styled } from '../../stitches.config'
 
 /**
  * Box component
