+++ src/components/Checkbox/Checkbox.tsx
@@ -1,8 +1,8 @@
 import { Indicator, Root } from '@radix-ui/react-checkbox'
 import type * as Polymorphic from '@radix-ui/react-polymorphic'
 import React, { forwardRef, ForwardRefExoticComponent } from 'react'
-import type { CSSProps, StitchesVariants } from 'stitches.config'
-import { CSS, styled } from 'stitches.config'
+import type { CSSProps, StitchesVariants } from '../../stitches.config'
+import { CSS, styled } from '../../stitches.config'
 import { ConditionalWrapper } from '../../utils'
 import { Check, CheckIndeterminate } from '../Icons'
 import { Label } from '../Label'
