+++ src/components/Chip/Chip.tsx
@@ -1,7 +1,7 @@
 import type * as Polymorphic from '@radix-ui/react-polymorphic'
 import React, { forwardRef } from 'react'
-import type { CSSProps, StitchesVariants } from 'stitches.config'
-import { styled } from 'stitches.config'
+import type { CSSProps, StitchesVariants } from '../../stitches.config'
+import { styled } from '../../stitches.config'
 import { IconButton } from '../IconButton'
 import { Close } from '../Icons'
