+++ src/components/ConfirmDialog/ConfirmDialog.tsx
@@ -11,8 +11,8 @@ import {
 import type * as Polymorphic from '@radix-ui/react-polymorphic'
 import { Slot } from '@radix-ui/react-slot'
 import React, { ComponentProps, FC, forwardRef } from 'react'
-import type { CSSProps } from 'stitches.config'
-import { CSS, styled } from 'stitches.config'
+import type { CSSProps } from '../../stitches.config'
+import { CSS, styled } from '../../stitches.config'
 import { Button } from '../Button'
 import { Heading } from '../Heading'
 import { overlayAnimationStyles, overlayStyles } from '../Overlay'
