+++ src/components/Container/Container.tsx
@@ -1,4 +1,4 @@
-import { styled } from 'stitches.config'
+import { styled } from '../../stitches.config'
 
 /**
  * Container component
