+++ src/components/Dialog/Dialog.tsx
@@ -2,8 +2,8 @@ import { Close, Content, Overlay, Root, Trigger } from '@radix-ui/react-dialog'
 import type * as Polymorphic from '@radix-ui/react-polymorphic'
 import { Slot } from '@radix-ui/react-slot'
 import React, { FC, forwardRef } from 'react'
-import type { CSSProps } from 'stitches.config'
-import { CSS, styled } from 'stitches.config'
+import type { CSSProps } from '../../stitches.config'
+import { CSS, styled } from '../../stitches.config'
 import { IconButton } from '../IconButton'
 import { Close as Icon } from '../Icons'
 import { overlayAnimationStyles, overlayStyles } from '../Overlay'
