+++ src/components/Drawer/Drawer.tsx
@@ -1,8 +1,8 @@
 import { Close, Content, Root } from '@radix-ui/react-dialog'
 import type * as Polymorphic from '@radix-ui/react-polymorphic'
 import React, { forwardRef } from 'react'
-import type { CSS, CSSProps, StitchesVariants } from 'stitches.config'
-import { keyframes, styled } from 'stitches.config'
+import type { CSS, CSSProps, StitchesVariants } from '../../stitches.config'
+import { keyframes, styled } from '../../stitches.config'
 import { DialogClose, DialogTrigger, StyledOverlay } from '../Dialog/Dialog'
 import { IconButton } from '../IconButton'
 import { Close as Icon } from '../Icons'
