+++ src/components/Grid/Grid.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Story, Meta } from '@storybook/react'
 import { Grid } from '.'
 import { Box, Row } from '../'
-import type { CSS } from 'stitches.config'
+import type { CSS } from '../../stitches.config'
 
 export default {
   title: 'Components/Grid',
