+++ src/components/Icons/Icons.tsx
@@ -8,7 +8,7 @@ import {
   mdiClose,
 } from '@mdi/js'
 import React, { ComponentProps, FC } from 'react'
-import type { CSS } from 'stitches.config'
+import type { CSS } from '../../stitches.config'
 import { Svg } from '../Svg'
 
 type IconType = FC<ComponentProps<typeof Svg>>
