+++ src/components/Label/Label.tsx
@@ -1,8 +1,8 @@
 import { Root } from '@radix-ui/react-label'
 import type * as Polymorphic from '@radix-ui/react-polymorphic'
 import React, { forwardRef } from 'react'
-import type { CSSProps, StitchesVariants } from 'stitches.config'
-import { styled } from 'stitches.config'
+import type { CSSProps, StitchesVariants } from '../../stitches.config'
+import { styled } from '../../stitches.config'
 import { StyledText } from '../Text/Text'
 
 const StyledLabel = styled(Root, {
