+++ src/components/List/ListItemIcon.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { styled } from 'stitches.config'
+import { styled } from '../../stitches.config'
 
 /**
  * ListItemIcon component
