+++ src/components/List/ListItemText.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { styled } from 'stitches.config'
+import { styled } from '../../stitches.config'
 import { Box } from '../Box'
 import { Text } from '../Text'
