+++ src/components/Logo/Logo.tsx
@@ -1,5 +1,5 @@
 import React, { ComponentProps, FC } from 'react'
-import type { CSS } from 'stitches.config'
+import type { CSS } from '../../stitches.config'
 import { Svg } from '../Svg'
 
 export const Logo: FC<ComponentProps<typeof Svg>> = ({ css, ...props }) => (
