+++ src/components/Paper/Paper.tsx
@@ -1,5 +1,5 @@
-import type { CSS } from 'stitches.config'
-import { styled } from 'stitches.config'
+import type { CSS } from '../../stitches.config'
+import { styled } from '../../stitches.config'
 
 export const paperStyles: CSS = {
   appearance: 'none',
