+++ src/components/Progress/Progress.tsx
@@ -1,7 +1,7 @@
 import type * as Polymorphic from '@radix-ui/react-polymorphic'
 import React, { forwardRef } from 'react'
-import type { CSSProps, StitchesVariants } from 'stitches.config'
-import { styled, keyframes } from 'stitches.config'
+import type { CSSProps, StitchesVariants } from '../../stitches.config'
+import { styled, keyframes } from '../../stitches.config'
 import { Root, Indicator } from '@radix-ui/react-progress'
 
 const indeterminateProgress = keyframes({
