+++ src/components/Table/Table.tsx
@@ -1,4 +1,4 @@
-import { styled } from 'stitches.config'
+import { styled } from '../../stitches.config'
 
 //------- Utility wrappers --------------//
