+++ src/components/Text/Text.tsx
@@ -1,7 +1,7 @@
 import type * as Polymorphic from '@radix-ui/react-polymorphic'
 import React, { forwardRef, PropsWithChildren } from 'react'
-import type { CSS, CSSProps, StitchesVariants } from 'stitches.config'
-import { styled } from 'stitches.config'
+import type { CSS, CSSProps, StitchesVariants } from '../../stitches.config'
+import { styled } from '../../stitches.config'
 
 const DEFAULT_TAG = 'span'
