+++ src/components/ThemeSwitch/ThemeSwitch.stories.tsx
@@ -6,7 +6,7 @@ import { Container } from '../Container'
 import { Heading } from '../Heading'
 import { Paper } from '../Paper'
 import { ThemeProvider } from '../ThemeProvider'
-import { lightTheme } from '../../utils/stitches.config'
+import { lightTheme } from '../../stitches.config'
 
 export default {
   title: 'Components/ThemeSwitch',
