+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from './components'
 
 // Stitches
-export * from './utils/stitches.config'
+export * from './stitches.config'
